requirements.txt
pika==1.3.2
load-dotenv==0.1.0
pika==1.3.2
python-dotenv==1.0.1
load-dotenv==0.1.0
pika==1.3.2
python-dotenv==1.0.1
